Make `literal!` no longer require `unsafe`
This is achieved by making the macro consume a string literal rather
than a byte string literal.
The issue being, we need a sized pointee to construct the static `ArcStr`.
For that, dereferencing the equivalent byte string literal solves the issue.

Saldy, there is no built-in way to convert a string literal into a byte
string literal, but it is something a procedural macro can easily do,
with **no dependencies whatsoever**.

Thus, this PR:

  - implements that procedural macro and its usage,

  - renames the macro to the more friendly `literal!` (based on the docs);

  - adapts the testing code accordingly;

  - such change **bumps the MSRV to 1.45.0** (or otherwise the proc-macro-hack
    dependency needs to be pulled; know that it is lightweight)

  - such change makes the macro **incompatible with the `include_***!` family of
    macros**;

      - The associated test has been commented out and `#[ignore]`-annotated,
        I'll let the author decide how we handle this